requirements.txt
# Python dependencies
Pillow>=10.0.0
torch==2.7.0
transformers>=4.30.0
PyYAML>=6.0
tqdm>=4.65.0
accelerate>=0.20.0
pdf2image>=1.16.0
addict>=2.4.0
matplotlib>=3.7.0
torchvision==0.22.0
easydict>=1.13
einops>=0.8.0
# flash-attn==2.7.4 - Install separately from wheel for Windows with CUDA 12.8

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
